+++ src/main/java/org/testes/Forma.java
@@ -1,20 +1,20 @@
 package org.testes;
 
 public class Forma {
-    int area;
+    double area;
 
-    public Forma(int area) {
+    public Forma(double area) {
         this.area = area;
     }
     public Forma() {
 
     }
 
-    public int getArea() {
+    public double getArea() {
         return area;
     }
 
-    public void setArea(int area) {
+    public void setArea(double area) {
         this.area = area;
     }
 }
